Summary: Create a basic CRUD Repository for every entity Issue: link: https://gitlab.com/kv.061.java/project_kv_061/issues/3 Changes: Added classes and repositories for entities (Address, GlobalProperties, Identity, IdentityType, PropertyType, Tenant, TenantDetails, TenantProperties, User and UserDetails)

+++ src/main/java/com/softserve/itacademy/kek/repositories/GlobalPropertiesRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface GlobalPropertiesRepository extends CrudRepository<GlobalProperties, Long> {
+public interface GlobalPropertiesRepository extends CrudRepository<GlobalProperties, Integer> {
 }
