+++ src/main/java/com/softserve/itacademy/kek/repositories/TenantPropertiesRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TenantPropertiesRepository extends CrudRepository<TenantProperties, Long> {
+public interface TenantPropertiesRepository extends CrudRepository<TenantProperties, Integer> {
 }
